Add a "type" field to workflow info dictionary
For alignment with a forthcoming dataclass implementation, and in anticipation of more interesting types like "While" and "For". As it is though, it doesn't really hurt and just makes explicit what sort of node you're looking at instead of needing to infer it from the presence of fields ("function" for "Function" types, "nodes", "edges" for "Workflow" types) -- i.e. at worst a little redundant.

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>